[fix]UI画像調整。キャラが動かなくなるバグ修正。敵がプレイヤーを検視しないバグ修正

+++ SirensMoon/Actor.cpp
@@ -45,6 +45,6 @@ void Actor::CheckDeadOwner() {
 
 void Actor::CheckRoomPosition() {
 	double x = floor((_pos.x + _size.x / 2.0) / (static_cast<double>(splitscreen_W)));
-	double y = floor((_pos.y + _size.y / 2) / screen_H);
+	double y = floor((_pos.y + _size.y / 2.0) / (static_cast<double>(screen_H)));
 	_roomPosition = { x,y };
 }
